requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
backoff==2.2.1
bitarray==3.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cytoolz==1.0.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.1.0
Flask==3.0.3
frozenlist==1.5.0
gql==3.5.0
graphql-core==3.2.5
gunicorn==23.0.0
hexbytes==1.2.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
plotly==5.24.1
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rlp==4.0.1
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
web3==7.6.1
websockets==13.1
Werkzeug==3.0.6
yarl==1.18.3
zipp==3.21.0
